Arreglos

+++ database/migrations/2019_09_07_161143_create_cabecera_movimientos_table.php
@@ -17,7 +17,7 @@ class CreateCabeceraMovimientosTable extends Migration
             $table->bigIncrements('id');
             $table->date('fecha');
             $table->date('fechaComprobante')->nullable();
-            $table->integer('numeroComprobante')->nullable();
+            $table->bigInteger('numeroComprobante')->nullable();
             $table->unsignedBigInteger('proveedor_id')->nullable();
             $table->foreign('proveedor_id')->references('id')->on('proveedores') ;
             $table->unsignedBigInteger('trabajo_id')->nullable();
